EA-108397 - assert: PsiSearchRequest.<init>
